+++ spec/Domain/User/UserIdSpec.php
@@ -2,6 +2,7 @@
 
 namespace spec\CubicMushroom\Hexagonal\Domain\User;
 
+use CubicMushroom\Hexagonal\Domain\Generic\ModelId;
 use CubicMushroom\Hexagonal\Domain\User\UserId;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
@@ -15,11 +16,29 @@ use Prophecy\Argument;
  */
 class UserIdSpec extends ObjectBehavior
 {
-    function it_is_initializable()
+    const NUMBER_ID = 123;
+
+
+    const STRING_ID = 'abc';
+
+
+    function let()
     {
         /** @noinspection PhpMethodParametersCountMismatchInspection */
-        $this->beConstructedWith(123);
+        $this->beConstructedWith(self::NUMBER_ID);
+    }
+
+
+    function it_is_initializable()
+    {
         $this->shouldHaveType(UserId::class);
+    }
+
+
+    function it_extends_model_id()
+    {
+        /** @noinspection PhpUndefinedMethodInspection */
+        $this->shouldBeAnInstanceOf(ModelId::class);
     }
 
 
@@ -28,24 +47,31 @@ class UserIdSpec extends ObjectBehavior
      */
     function it_can_be_initialised_with_an_integer()
     {
-        $id = 123;
 
         /** @noinspection PhpMethodParametersCountMismatchInspection */
-        $this->beConstructedWith($id);
+        $this->beConstructedWith(self::NUMBER_ID);
 
         /** @noinspection PhpUndefinedMethodInspection */
-        $this->getId()->shouldReturn($id);
+        $this->getId()->shouldReturn(self::NUMBER_ID);
     }
 
 
     function it_can_be_initialised_with_a_string()
     {
-        $id = 'abc';
-
         /** @noinspection PhpMethodParametersCountMismatchInspection */
-        $this->beConstructedWith($id);
+        $this->beConstructedWith(self::STRING_ID);
 
         /** @noinspection PhpUndefinedMethodInspection */
-        $this->getId()->shouldReturn($id);
+        $this->getId()->shouldReturn(self::STRING_ID);
+    }
+
+
+    function it_can_be_converted_to_a_string()
+    {
+        /** @noinspection PhpMethodParametersCountMismatchInspection */
+        $this->beConstructedWith(self::NUMBER_ID);
+
+        /** @noinspection PhpUndefinedMethodInspection */
+        $this->__toString()->shouldReturn((string) self::NUMBER_ID);
     }
 }