+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## v1.0.3 - 2015-10-18
 ### Added
 - Logging exceptions if validator is not available to validate command
-- ModelID can be converted to string
+- ModelID/UserId can be converted to string
 ### Fixed
 - n/a
 ### Changed